LPS-73924 Portlet Split: extract base class from workflow portlet, to reuse in different sections of left menu.
